task1/requirements.txt
Django==3.2.15
django-nested-admin==4.0.2
psycopg2-binary>=2.8